Update the font size of the footer text
As a general rule of thumb, I'm making sure that the font size is always
16px or higher. There's a weird bug that the site zooms in slightly on
mobile devices despite width=device-width and initial-scale=1. One
theory is that there's an input field with font-size < 16px, but there's
none that I can find. Only the footer text is smaller than 16px, so I'm
fixing it to see if that helps (probably won't, but worth a shot).