Enable task addition from agenda page
On three different occasions, I found myself pressing 'n' to add a task
while on the agenda page (which of course didn't work). So it seemed
like a good feature to have. I have intentionally NOT added new tasks to
the agenda automatically, because it's a good habit not to preempt the
existing agenda.