Update behavior to make items clickable in list
I decided to make the behavior consistent across all items. Previously,
tasks and tags could be edited on click, whereas notes would like to the
detail page. Tasks on the agenda page would, on the other hand, be
clickable. With this change, *everything* is clickable to get to its
detail page. You can edit once you get to the detail page.