Update styling for scheduled tasks / works.
* Keeps scheduled tasks in the list (no longer moves to bottom).
* Reverts to standard styling.
* Replaces the 'today' symbol with an 'alarm on' symbol (if scheduled).

THe motivation for this change is that it was odd, and a bit annoying, to
find tasks that had been scheduled, as they tended to disappear down the
list. Also, this was inconsistent with the 'add to agenda' button
behavior where it remained in the same place in the list.