Make back button lead back to search results page
Fix: https://github.com/rri/dinky/issues/3

When the user enters a search term, then clicks on an item to navigate
to its details, the corresponding back button originally wouldn't lead
the user back to the search results (because that page wasn't in the
browser's history).

To fix this, we do several things:

* Ensure that when the search term is updated, an encoded version of it
  gets updated in the browser's URL hash as well. Without any additional
  work, hitting the back button leads the user back to this URL (with
  the hash appended).
* Ensure that whenever a page loads with the URL hash set to something,
  treat it as a search term by updating the corresponding state.

The rest, as they say, is history.