Remove agenda customization
This commit removes the ability for users to customize morning and
evening buffer hours. The main change in behavior this introduces is a
simplification in how the agenda works. Instead of showing items on the
date it is tagged with, the agenda for tasks shows items on the date it
is tagged with AND all future dates. Effectively, this make the agenda
for tasks behave the same way as the library: items remain sticky until
removed or completed.